backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
langchain==0.1.0
langchain-google-genai==0.0.6
langgraph==0.0.26
qdrant-client==1.7.0
google-generativeai==0.3.2
python-multipart==0.0.6
httpx==0.25.2
